Move unix_op to unix::op.
